fix: Support restricting instance type in `update-instance-type` CLI
Currently, `update-instance-type` CLI only updates when there are
new instnace types in the local config files.

This change adds support for restricting instance types if the
local config file has missing types that are in the existing
listing.

If there are no additional or missing instance types, no changeset
will be generated and the request will fail.