feat: Support AmiProduct download CLI
Add a `download` CLI command to get configuration
files from listings in AWS Marketplace.

The downloaded configuration files have a same
foramt as the local YAML config file. This can
be useful for users who do not have a the full
configuration but request a changes to an existing
listing.

The LiteralString class is introduced to dump strings
in literal block style(|) in the YAML file.